+++ odoo/local-src/sale_cancel_remaining/__manifest__.py
@@ -13,6 +13,7 @@
     """,
     'depends': [
         'sale',
+        'stock',
     ],
     'data': [
         "wizard/cancel_remaining_wizard.xml",